+++ gazebo/gui/ConfigWidget.hh
@@ -40,6 +40,8 @@ namespace gazebo
 {
   namespace gui
   {
+    class GroupWidget;
+
     /// \addtogroup gazebo_gui
     /// \{
 
@@ -50,8 +52,15 @@ namespace gazebo
     {
       Q_OBJECT
 
+      /// \brief Constructor;
+      public: ConfigChildWidget() : groupWidget(NULL) {}
+
       /// \brief List of child widgets.
       public: std::vector<QWidget *> widgets;
+
+      /// \brief Pointer to group widget.
+      /// NULL if this widget is not contained inside a group widget.
+      public: GroupWidget *groupWidget;
     };
 
     /// \class GeometryConfigWidget ConfigWidget.hh
@@ -100,82 +109,151 @@ namespace gazebo
       /// \brief Destructor
       public: ~ConfigWidget();
 
-      /// \brief Load from a google protobuf message
-      /// \param[in] _msg Message to load from
+      /// \brief Load from a google protobuf message.
+      /// \param[in] _msg Message to load from.
       public: void Load(const google::protobuf::Message *_msg);
 
       /// \brief Get the updated message.
       /// \return Updated message.
       public: google::protobuf::Message *GetMsg();
 
+      /// \brief Set whether a child widget should be visible.
+      /// \param[in] _name Name of the child widget.
+      /// \param[in] _visible True to set the widget to be visible.
+      public: void SetWidgetVisible(const std::string &_name, bool _visible);
+
+      /// \brief Get whether a child widget is visible.
+      /// \param[in] _name Name of the child widget.
+      /// \return True if the widget is visible.
+      public: bool GetWidgetVisible(const std::string &_name) const;
+
+      /// \brief Set whether a child widget should be read-only.
+      /// \param[in] _name Name of the child widget.
+      /// \param[in] _visible True to set the widget to be read-only.
+      public: void SetWidgetReadOnly(const std::string &_name, bool _readOnly);
+
+      /// \brief Get whether a child widget is read-only.
+      /// \param[in] _name Name of the child widget.
+      /// \return True if the widget is read-only.
+      public: bool GetWidgetReadOnly(const std::string &_name) const;
+
       /// \brief Update the widgets from a message.
       /// \param[in] _msg Message used for updating the widgets.
       public: void UpdateFromMsg(const google::protobuf::Message *_msg);
 
-      /// \brief Set an integer value to a child wiget in the config widget.
-      /// \param[in] _name Name of the widget.
+      /// \brief Set an integer value to a child widget.
+      /// \param[in] _name Name of the child widget.
       /// \param[in] _value Value to set to.
-      public: void SetIntWidgetProperty(const std::string &_name, int _value);
+      public: void SetIntWidgetValue(const std::string &_name, int _value);
 
-      /// \brief Set an unsigned integer value to a child wiget in the
-      /// config widget.
-      /// \param[in] _name Name of the widget.
+      /// \brief Set an unsigned integer value to a child widget.
+      /// \param[in] _name Name of the child widget.
       /// \param[in] _value Value to set to.
-      public: void SetUIntWidgetProperty(const std::string &_name, unsigned int
+      public: void SetUIntWidgetValue(const std::string &_name, unsigned int
           _value);
 
-      /// \brief Set a double value to a child wiget in the config widget.
-      /// \param[in] _name Name of the widget.
+      /// \brief Set a double value to a child widget.
+      /// \param[in] _name Name of the child widget.
       /// \param[in] _value Value to set to.
-      public: void SetDoubleWidgetProperty(const std::string &_name,
+      public: void SetDoubleWidgetValue(const std::string &_name,
           double _value);
 
-      /// \brief Set a bool value to a child wiget in the config widget.
-      /// \param[in] _name Name of the widget.
+      /// \brief Set a bool value to a child widget.
+      /// \param[in] _name Name of the child widget.
       /// \param[in] _value Value to set to.
-      public: void SetBoolWidgetProperty(const std::string &_name, bool _value);
+      public: void SetBoolWidgetValue(const std::string &_name, bool _value);
 
-      /// \brief Set a string value to a child wiget in the config widget.
-      /// \param[in] _name Name of the widget.
+      /// \brief Set a string value to a child widget.
+      /// \param[in] _name Name of the child widget.
       /// \param[in] _value Value to set to.
-      public: void SetStringWidgetProperty(const std::string &_name,
+      public: void SetStringWidgetValue(const std::string &_name,
           const std::string &_value);
 
-      /// \brief Set a vector3 value to a child wiget in the config widget.
-      /// \param[in] _name Name of the widget.
+      /// \brief Set a vector3 value to a child widget.
+      /// \param[in] _name Name of the child widget.
       /// \param[in] _value Value to set to.
-      public: void SetVector3WidgetProperty(const std::string &_name,
+      public: void SetVector3WidgetValue(const std::string &_name,
           const math::Vector3 &_value);
 
-      /// \brief Set a color value to a child wiget in the config widget.
-      /// \param[in] _name Name of the widget.
+      /// \brief Set a color value to a child widget.
+      /// \param[in] _name Name of the child widget.
       /// \param[in] _value Value to set to.
-      public: void SetColorWidgetProperty(const std::string &_name,
+      public: void SetColorWidgetValue(const std::string &_name,
           const common::Color &_value);
 
-      /// \brief Set a pose value to a child wiget in the config widget.
-      /// \param[in] _name Name of the widget.
+      /// \brief Set a pose value to a child widget.
+      /// \param[in] _name Name of the child widget.
       /// \param[in] _value Value to set to.
-      public: void SetPoseWidgetProperty(const std::string &_name,
+      public: void SetPoseWidgetValue(const std::string &_name,
           const math::Pose &_value);
 
-      /// \brief Set a geometry value to a child wiget in the config widget.
-      /// \param[in] _name Name of the widget.
+      /// \brief Set a geometry value to a child widget.
+      /// \param[in] _name Name of the child widget.
       /// \param[in] _value Type of geometry.
       /// \param[in] _dimensions Dimensions of geometry.
-      public: void SetGeometryWidgetProperty(const std::string &_name,
+      public: void SetGeometryWidgetValue(const std::string &_name,
           const std::string &_value, const math::Vector3 &_dimensions);
+
+      /// \brief Get an integer value from a child widget.
+      /// \param[in] _name Name of the child widget.
+      /// \return Integer value.
+      public: int GetIntWidgetValue(const std::string &_name) const;
+
+      /// \brief Get an unsigned integer value from a child widget.
+      /// \param[in] _name Name of the child widget.
+      /// \return Unsigned integer value.
+      public: unsigned int GetUIntWidgetValue(const std::string &_name) const;
+
+      /// \brief Get a double value from a child widget.
+      /// \param[in] _name Name of the child widget.
+      /// \return Double value.
+      public: double GetDoubleWidgetValue(const std::string &_name) const;
+
+      /// \brief Get a bool value from a child widget.
+      /// \param[in] _name Name of the child widget.
+      /// \return Bool value.
+      public: bool GetBoolWidgetValue(const std::string &_name) const;
+
+      /// \brief Get a string value from a child widget.
+      /// \param[in] _name Name of the child widget.
+      /// \return String value.
+      public: std::string GetStringWidgetValue(const std::string &_name) const;
+
+      /// \brief Get a vector3 value from a child widget.
+      /// \param[in] _name Name of the child widget.
+      /// \return Vector3 value.
+      public: math::Vector3 GetVector3WidgetValue(const std::string &_name)
+          const;
+
+      /// \brief Get a color value from a child widget.
+      /// \param[in] _name Name of the child widget.
+      /// \return Color value.
+      public: common::Color GetColorWidgetValue(const std::string &_name) const;
+
+      /// \brief Get a pose value from a child widget.
+      /// \param[in] _name Name of the child widget.
+      /// \return Pose value.
+      public: math::Pose GetPoseWidgetValue(const std::string &_name) const;
+
+      /// \brief Get a geometry value from a child widget.
+      /// \param[in] _name Name of the child widget.
+      /// \param[out] _dimensions Dimensions of geometry.
+      /// \return Type of geometry.
+      public: std::string GetGeometryWidgetValue(const std::string &_name,
+          math::Vector3 &_dimensions) const;
 
       /// \brief Parse the input message and either create widgets for
       /// configuring fields of the message, or update the widgets with values
       /// from the message.
+      /// \param[in] _msg Message.
       /// \param[in] _name Name used when creating new widgets.
+      /// return Updated widget.
       private: QWidget *Parse(google::protobuf::Message *_msg,
           const std::string &_name = "");
 
       /// \brief Parse a vector3 message.
       /// param[in] _msg Input vector3d message.
-      /// return Parsed vector
+      /// return Parsed vector.
       private: math::Vector3 ParseVector3(
           const google::protobuf::Message *_msg);
 
@@ -189,7 +267,7 @@ namespace gazebo
       /// \param[in] _msg Vector3d message to be updated.
       /// \param[in] _value Vector3 used for updating the message.
       private: void UpdateVector3Msg(google::protobuf::Message *_msg,
-          const math::Vector3 _value);
+          const math::Vector3 &_value);
 
       /// \brief Create a widget for configuring an unsigned integer value.
       /// \param[in] _key A key that is used as a label for the widget.
@@ -236,58 +314,109 @@ namespace gazebo
       /// \return The newly created widget.
       private: ConfigChildWidget *CreateGeometryWidget(const std::string &_key);
 
-      /// \brief Update the input widget with an unsigned integer value.
-      /// \param[in] _widget Pointer to the widget.
+      /// \brief Update a child widget with an unsigned integer value.
+      /// \param[in] _widget Pointer to the child widget.
       /// \param[in] _value Value to set to.
       private: void UpdateUIntWidget(ConfigChildWidget *_widget,
           unsigned int _value);
 
-      /// \brief Update the input widget with an integer value.
-      /// \param[in] _widget Pointer to the widget.
+      /// \brief Update a child widget with an integer value.
+      /// \param[in] _widget Pointer to the child widget.
       /// \param[in] _value Value to set to.
       private: void UpdateIntWidget(ConfigChildWidget *_widget, int _value);
 
-      /// \brief Update the input widget with a double value.
-      /// \param[in] _widget Pointer to the widget.
+      /// \brief Update a child widget with a double value.
+      /// \param[in] _widget Pointer to the child widget.
       /// \param[in] _value Value to set to.
       private: void UpdateDoubleWidget(ConfigChildWidget *_widget,
           double _value);
 
-      /// \brief Update the input widget with a string value.
-      /// \param[in] _widget Pointer to the widget.
+      /// \brief Update a child widget with a string value.
+      /// \param[in] _widget Pointer to the child widget.
       /// \param[in] _value Value to set to.
       private: void UpdateStringWidget(ConfigChildWidget *_widget,
           const std::string &_value);
 
-      /// \brief Update the input widget with a bool value.
-      /// \param[in] _widget Pointer to the widget.
+      /// \brief Update a child widget with a bool value.
+      /// \param[in] _widget Pointer to the child widget.
       /// \param[in] _value Value to set to.
       private: void UpdateBoolWidget(ConfigChildWidget *_widget, bool _value);
 
-      /// \brief Update the input widget with a vector3 value.
-      /// \param[in] _widget Pointer to the widget.
+      /// \brief Update a child widget with a vector3 value.
+      /// \param[in] _widget Pointer to the child widget.
       /// \param[in] _value Value to set to.
       private: void UpdateVector3Widget(ConfigChildWidget *_widget,
           const math::Vector3 &_value);
 
-      /// \brief Update the input widget with a color value.
-      /// \param[in] _widget Pointer to the widget.
+      /// \brief Update a child widget with a color value.
+      /// \param[in] _widget Pointer to the child widget.
       /// \param[in] _value Value to set to.
       private: void UpdateColorWidget(ConfigChildWidget *_widget,
           const common::Color &_value);
 
-      /// \brief Update the input widget with a pose value.
-      /// \param[in] _widget Pointer to the widget.
+      /// \brief Update a child widget with a pose value.
+      /// \param[in] _widget Pointer to the child widget.
       /// \param[in] _value Value to set to.
       private: void UpdatePoseWidget(ConfigChildWidget *_widget,
           const math::Pose &_value);
 
-      /// \brief Update the input widget with a geometry type and dimensions.
-      /// \param[in] _widget Pointer to the widget.
-      /// \param[in] _value Value Type of geometry
+      /// \brief Update a child widget with a geometry type and dimensions.
+      /// \param[in] _widget Pointer to the child widget.
+      /// \param[in] _value Type of geometry.
       /// \param[in] _dimensions Dimensions of the geometry.
       private: void UpdateGeometryWidget(ConfigChildWidget *_widget,
           const std::string &_value, const math::Vector3 &_dimensions);
+
+      /// \brief Get an integer value from a child widget.
+      /// \param[in] _widget Pointer to the child widget.
+      /// \return Value of the widget.
+      private: int GetIntWidgetValue(ConfigChildWidget *_widget) const;
+
+      /// \brief Get an unsigned integer value from a child widget.
+      /// \param[in] _widget Pointer to the child widget.
+      /// \return Value of the widget.
+      private: unsigned int GetUIntWidgetValue(ConfigChildWidget *_widget)
+          const;
+
+      /// \brief Get a double value from a child widget.
+      /// \param[in] _widget Pointer to the child widget.
+      /// \return Value of the widget.
+      private: double GetDoubleWidgetValue(ConfigChildWidget *_widget) const;
+
+      /// \brief Get a bool value from a child widget.
+      /// \param[in] _widget Pointer to the child widget.
+      /// \return Value of the widget.
+      private: bool GetBoolWidgetValue(ConfigChildWidget *_widget) const;
+
+      /// \brief Get a string value from a child widget.
+      /// \param[in] _widget Pointer to the child widget.
+      /// \return Value of the widget.
+      private: std::string GetStringWidgetValue(ConfigChildWidget *_widget)
+          const;
+
+      /// \brief Get a vector3 value from a child widget.
+      /// \param[in] _widget Pointer to the child widget.
+      /// \return Value of the widget.
+      private: math::Vector3 GetVector3WidgetValue(ConfigChildWidget *_widget)
+          const;
+
+      /// \brief Get a color value from a child widget.
+      /// \param[in] _widget Pointer to the child widget.
+      /// \return Value of the widget.
+      private: common::Color GetColorWidgetValue(ConfigChildWidget *_widget)
+          const;
+
+      /// \brief Get a pose value from a child widget.
+      /// \param[in] _widget Pointer to the child widget.
+      /// \return Value of the widget.
+      private: math::Pose GetPoseWidgetValue(ConfigChildWidget *_widget) const;
+
+      /// \brief Get a geometry value from a child widget.
+      /// \param[in] _widget Pointer to the child widget.
+      /// \param[out] _dimensions Dimensions of geometry.
+      /// \return Type of geometry.
+      private: std::string GetGeometryWidgetValue(ConfigChildWidget *_widget,
+          math::Vector3 &_dimensions) const;
 
       /// \brief Received item selection user input.
       /// \param[in] _item Item selected.